refactor: remove weather feature
